Create Leaderships when creating a group
Summary:
This commit allows for leadership relations to be created when a group
is created. It adds server support for leadership via nested attributes
and client side changes.

Test Plan:
Navigate to a conference, create a new group, and you should
see leaders as options when creating this group.

Reviewers: warren

Reviewed By: warren

Subscribers: a.h, qromanek, sonia, unzunz

Projects: #edge-yl

Differential Revision: http://phab.calblueprint.org/D169