Auto-associate emails with models and mark as read
Summary:
Now when an email is received with a create request, it
will automatically associate itself with the correct emailable and user
based on what the sender and recipient emails are. Also fixed the
validation method in email.rb so that it only will send an email that is
being sent by the user.

Test Plan: Ha ha

Reviewers: a.h, trickywolffe, unzunz, sonia, warren

Reviewed By: warren

Subscribers: sonia, unzunz, trickywolffe, a.h

Projects: #edge-yl

Differential Revision: http://phab.calblueprint.org/D219