Refactored email components and controllers
Summary:
Now whether a draft or email is created, you only need to send the
sender and recipient emails and the set_initials method will
auto-generate the model relations by checking the database for
emailables/users with those emails.  The from/to fields will also be
updated automatically.

Now when typing quickly the text will not be reset when the save request
completes.

While waiting for a draft to be saved "Saving..." is displayed

Test Plan: Hopefully some tests

Reviewers: a.h, trickywolffe, unzunz, sonia, warren

Reviewed By: warren

Subscribers: sonia, unzunz, trickywolffe, a.h

Projects: #edge-yl

Differential Revision: http://phab.calblueprint.org/D252