Groups - change URL based on selected Conference
Summary:
Fix the Groups index view to change the url based on the selected
conference. The state of the url is also stored, so that when the back
button is pressed, the correct set of groups, and conference is
re-inserted.

Fix #1084

Test Plan:
Go to groups and fiddle with the conference drop down, you
should see the conference change, url change, and set of groups
change. Now go back with your browser, the history should be
preserved.

Reviewers: a.h, qromanek, sonia, unzunz, warren

Reviewed By: warren

Subscribers: a.h, qromanek, sonia, unzunz

Projects: #edge-yl

Differential Revision: http://phab.calblueprint.org/D276