Update rooms page URL when filtering to a new conference
Summary: Fix the rooms index page to update its URL when selecting a new conference. This is done by pushing the url's state when selecting a new state and popping it when pressing the back button.

Test Plan: Go to the rooms index view and select a new conference from the drop down. You should see the conference name and url change. When pressing the back button, confirm that the previous list of rooms and conference are rendered.

Reviewers: a.h, qromanek, trickywolffe, sonia, warren

Reviewed By: warren

Subscribers: a.h, qromanek, trickywolffe, sonia

Projects: #edge-yl

Differential Revision: http://phab.calblueprint.org/D280