Fix Dispatch Error bug in AddStudent feature
Summary:
Fix #1149

I was accidentally firing two actions when the save button was pressed
when adding a student to a group. If the focus was still on the search
input field when the save button was pressed, then handleBlur got called
which fired an action. I went around in circles looking for the bug,
because it still appeared even after removing handleBlur (because the
searchInput handle blur was being called instead.) Moral of the story -
don't be dumb, and be careful about JS inheritance, because there's not
an easy way to tell where dispatch errors are coming from in alt.

Test Plan:
Open console and add a student with the focus on the input,
no dispatch error. Woot.

Reviewers: warren, a.h, qromanek, sonia, unzunz

Reviewed By: warren

Subscribers: a.h, qromanek, sonia, unzunz

Projects: #edge-yl

Differential Revision: http://phab.calblueprint.org/D304