Fix Conference and Group Bugs
Summary:
Fix #1233
Fix #1234
Fix #1236

Add an alert when assignStudentsToRooms is called without enough
room space. Fix a bug which allowed students to be added to a group that
was not in the same conference as the student. Fix a small propType bug.

Test Plan:
Try to add students to a group within a conference which has
no students, you should see no results.

Reviewers: unzunz, sonia, qromanek, a.h, warren

Reviewed By: warren

Subscribers: a.h, qromanek, sonia, unzunz

Projects: #edge-yl

Differential Revision: http://phab.calblueprint.org/D331