Remove Rooms and Groups index view.
Summary:
We're moving to a more conference based approach, this commit removes
the index pages of rooms and groups, which will instead be visible from
the conference show view.

Fix #1243

Test Plan:
You should not see any references to the rooms page or the
groups page, and both routes should not work.

Reviewers: unzunz, sonia, qromanek, a.h, warren

Reviewed By: warren

Subscribers: a.h, qromanek, sonia, unzunz

Projects: #edge-yl

Differential Revision: http://phab.calblueprint.org/D341