Delete Rooms and Groups from Conference seed update
Summary:
This commit adds proper delete functionality to remove rooms
and groups from conference. It also adds not null constraints to
rooms, groups, and students, to ensure that they always have a
conference. Finally, it fixes the seeds files to only assign
students to groups and rooms for a single conference

Test Plan: Naw

Reviewers: unzunz, sonia, qromanek, a.h, warren

Reviewed By: warren

Subscribers: a.h, qromanek, sonia, unzunz

Projects: #edge-yl

Differential Revision: http://phab.calblueprint.org/D388