Add Check In conference related features
Summary:
Fix #1408
Fix #1447
Fix #1420

This commit adds the ability to change conferences from the check in
screen, as well as some information about the conference which is
currently accepting check-ins, such as number checked in for each
gender, and total number of checked in students

Test Plan: Nope. One day.

Reviewers: sonia, qromanek, a.h, warren

Reviewed By: warren

Subscribers: a.h, qromanek, sonia, unzunz

Projects: #edge-yl

Differential Revision: http://phab.calblueprint.org/D441