Set up responsibilities
Summary: Create Responsibility model. Create a callback for conference model to create responsibilities for each school when a conference is created. Render responsibilities of a user on user show page. Include the name of the conference in Student Conference information card

Test Plan: Create a new conference and make sure responsibilities are created for each school.

Reviewers: warren

Reviewed By: warren

Subscribers: a.h, qromanek, trickywolffe, sonia

Projects: #edge-yl

Differential Revision: http://phab.calblueprint.org/D449