Assign user responsibility from school view page
Summary: Create ResponsibilityEditModal component to assign user to responsibility. The component displays a dropdown a list of all unassigned users.  From this list, a user can assign a responsibility a user.

Test Plan: Go to a school view page, click edit, assign a responsibility, and confirm that the udpated responsibility renders.

Reviewers: warren

Reviewed By: warren

Subscribers: a.h, qromanek, trickywolffe, sonia

Projects: #edge-yl

Differential Revision: http://phab.calblueprint.org/D459