Make rooms binary gendered, fix remove student.
Summary:
This commit makes room genders either male or female. Students must have
a male or female gender to be automatically assigned to a room by the
conference. It also fixes the ability to remove students from rooms. It
also allows a user to modify the gender of a room.

Fix #1550
Fix #1495

Test Plan: Manual :(

Reviewers: unzunz, a.h, qromanek, sonia, warren

Reviewed By: warren

Subscribers: a.h, qromanek, sonia, unzunz

Projects: #edge-yl

Differential Revision: http://phab.calblueprint.org/D543